refactor: presentation part 코드 구조 변경
