+++ src/main/java/com/ninedocs/userserver/user/presentation/authentication/EmailVerificationCodeController.java
@@ -1,4 +1,4 @@
-package com.ninedocs.userserver.user.presentation;
+package com.ninedocs.userserver.user.presentation.authentication;
 
 import com.ninedocs.userserver.user.application.emailverificationcode.EmailVerificationCodeService;
 import com.ninedocs.userserver.user.application.emailverificationcode.dto.EmailDuplicateCodeRequest;
