+++ src/main/java/com/ninedocs/userserver/user/presentation/authentication/EmailVerificationController.java
@@ -1,4 +1,4 @@
-package com.ninedocs.userserver.user.presentation;
+package com.ninedocs.userserver.user.presentation.authentication;
 
 
 import com.ninedocs.userserver.common.presentation.dto.ApiResponse;
