+++ src/main/java/com/ninedocs/userserver/user/presentation/usermanagement/SignUpController.java
@@ -1,20 +1,12 @@
-package com.ninedocs.userserver.user.presentation;
+package com.ninedocs.userserver.user.presentation.usermanagement;
 
 import com.ninedocs.userserver.common.presentation.dto.ApiResponse;
-import com.ninedocs.userserver.user.application.emailverificationcode.EmailFormatValidator;
-import com.ninedocs.userserver.user.application.emailverificationcode.exception.EmailFormatException;
 import com.ninedocs.userserver.user.application.signin.JwtProvider;
 import com.ninedocs.userserver.user.application.signin.dto.JwtTokenResult;
-import com.ninedocs.userserver.user.application.signup.EmailVerificationChecker;
 import com.ninedocs.userserver.user.application.signup.SignUpService;
 import com.ninedocs.userserver.user.application.signup.dto.SignUpRequest;
 import com.ninedocs.userserver.user.application.signup.dto.SignUpResponse;
-import com.ninedocs.userserver.user.application.signup.exception.EmailNotVerifiedException;
-import com.ninedocs.userserver.user.persistence.User;
-import com.ninedocs.userserver.user.persistence.UserRepository;
 import jakarta.validation.Valid;
-import java.time.LocalDateTime;
-import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
